Move frame data out of vectorize interface land
