+++ tcn_hpl/data/vectorize/__init__.py
@@ -1,6 +1,1 @@
-from ._data import (
-    FrameObjectDetections,
-    FramePoses,
-    FrameData,
-)
 from ._interface import Vectorize
